requirements.txt
mysql-connector-python==8.0.33
pandas==2.0.3
matplotlib==3.7.2